requirements.txt
python-dotenv~=1.0.1
langchain~=0.2.3

requests~=2.32.3
